[DebuggerV2] Remove transitive dependency on store-internal helper
