+++ tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/store/debugger_store_utils.ts
@@ -16,8 +16,9 @@ limitations under the License.
  * Utility functions for the NgRx store of Debugger V2.
  */
 
+import {getFocusedStackFramesHelper} from './debugger_store_helpers';
+
 import {
-  CodeLocationType,
   DebuggerState,
   SourceFileSpec,
   SourceLineSpec,
@@ -157,53 +158,6 @@ export function beginEndRangesInclude(
   return (
     ranges.findIndex((range) => range.begin >= begin && range.end <= end) !== -1
   );
-}
-
-/**
- * Helper function that extracts the stack trace being focused on.
- *
- * This examines whether the current focused code location is for an
- * eager (top-level) execution or a graph-op creation, and then queries
- * the corresponding substates accordingly.
- *
- * @param state
- */
-export function getFocusedStackFramesHelper(
-  state: DebuggerState
-): StackFrame[] | null {
-  if (state.codeLocationFocusType === null) {
-    return null;
-  }
-  let stackFrameIds: string[] = [];
-  if (state.codeLocationFocusType === CodeLocationType.EXECUTION) {
-    const {focusIndex, executionData} = state.executions;
-    if (focusIndex === null || executionData[focusIndex] === undefined) {
-      return null;
-    }
-    stackFrameIds = executionData[focusIndex].stack_frame_ids;
-  } else {
-    // This is CodeLocationType.GRAPH_OP_CREATION.
-    if (state.graphs.focusedOp === null) {
-      return null;
-    }
-    const {graphId, opName} = state.graphs.focusedOp;
-    if (
-      state.graphs.ops[graphId] === undefined ||
-      !state.graphs.ops[graphId].has(opName)
-    ) {
-      return null;
-    }
-    stackFrameIds = state.graphs.ops[graphId].get(opName)!.stack_frame_ids;
-  }
-  const stackFrames: StackFrame[] = [];
-  for (const stackFrameId of stackFrameIds) {
-    if (state.stackFrames[stackFrameId] != null) {
-      stackFrames.push(state.stackFrames[stackFrameId]);
-    } else {
-      return null;
-    }
-  }
-  return stackFrames;
 }
 
 /**
